--- app/src/main/java/com/survivalcoding/maskinfo/TODO_LIST.kt
@@ -1,5 +1,0 @@
-package com.survivalcoding.maskinfo
-
-// presentation layer에서 권한 체크 로직 분리
-// data layer에서 Store.distanceTo use_case로 분리
-// Dagger 의존성 주입
